refactor: quiz app page.tsx로 하위 섹션 이동
